tests/requirements.txt
pytest==4.4.1
pytest-django==3.4.8
coolname==1.1.0
factory-boy==2.11.1
